delete LoginComponent, rename ButtonComp
