--- src/ui/components/ButtonPage/Button/ButtonContainer.js
@@ -1,15 +1,0 @@
-import React, { Component } from "react";
-import ButtonComponent from "./ButtonComponent";
-// Class HOC, which returns modificated component
-
-class ButtonContainer extends Component {
-  constructor(props) {
-    super(props);
-  }
-
-  render() {
-    return <ButtonComponent />;
-  }
-}
-
-export default ButtonContainer;
