--- src/ui/components/ButtonPage/Button/index.js
@@ -1,4 +1,0 @@
-import ButtonContainer from "./ButtonContainer";
-import "./ButtonComponent.css";
-
-export default ButtonContainer;
