--- src/ui/components/LoginPage/index.js
@@ -1,3 +1,0 @@
-import LoginPageContainer from './LoginPageContainer'
-
-export default LoginPageContainer;
