+++ src/ui/components/RandomizeScreen/RandomizeScreenComponent.scss
