+++ src/ui/components/RandomizeScreen/index.js
@@ -1,0 +1,4 @@
+import ButtonContainer from './RandomizeScreenContainer';
+import './ButtonComponent.scss';
+
+export default ButtonContainer;
